+++ tests/testthat/test-coef_rename.R
@@ -4,16 +4,6 @@ mod <- list()
 mod$OLS <- lm(am ~ drat, data = mtcars)
 mod$Logit <- glm(am ~ qsec, data = mtcars, family = binomial())
 
-test_that("unique names", {
-  cmap <- c('drat'='Constant', 'drat'='Rear axle ratio')
-  expect_error(modelsummary(mod, "data.frame", coef_rename = cmap),
-               regexp = "duplicated")
-
-  cmap <- c('Constant', 'Rear axle ratio')
-  expect_error(modelsummary(mod, "data.frame", coef_rename = cmap),
-               regexp = "Must have names")
-})
-
 test_that("rename 2 out of 3 coefficients", {
   cmap <- c('(Intercept)'='Constant', 'drat'='Rear axle ratio')
   mod <- modelsummary(mod, "dataframe", coef_rename=cmap)
@@ -34,6 +24,12 @@ test_that("regression test: coef_rename() function", {
     expect_error(modelsummary(dvnames(x), coef_rename = coef_rename), NA)
 })
 
+
+test_that("unnamed coef_rename + coef_omit numeric", {
+  mod <- lm(mpg ~ hp + wt + disp + carb + gear + vs + cyl + drat + am + hp * wt, data = mtcars)
+  tab <- modelsummary(mod, output = "data.frame", coef_omit = c(3:11), coef_rename = c("Constant", "Horsepower"), gof_map = NA)
+  expect_equal(nrow(tab), 4)
+})
 
 
 # Q